Settings MiscTab refactor
