+++ src/main/java/org/dpsoftware/gui/SettingsController.java
@@ -26,14 +26,12 @@ import javafx.application.Platform;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.Node;
 import javafx.scene.control.*;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.InputEvent;
-import javafx.scene.paint.Color;
 import javafx.stage.Stage;
 import javafx.util.Callback;
 import javafx.util.Duration;
@@ -42,25 +40,17 @@ import org.dpsoftware.FireflyLuciferin;
 import org.dpsoftware.JavaFXStarter;
 import org.dpsoftware.LEDCoordinate;
 import org.dpsoftware.NativeExecutor;
-import org.dpsoftware.audio.AudioLoopback;
-import org.dpsoftware.audio.AudioLoopbackSoftware;
-import org.dpsoftware.audio.AudioUtility;
 import org.dpsoftware.config.Configuration;
 import org.dpsoftware.config.Constants;
 import org.dpsoftware.gui.elements.DisplayInfo;
 import org.dpsoftware.managers.DisplayManager;
 import org.dpsoftware.managers.MQTTManager;
 import org.dpsoftware.managers.StorageManager;
-import org.dpsoftware.managers.dto.ColorDto;
 import org.dpsoftware.managers.dto.FirmwareConfigDto;
-import org.dpsoftware.managers.dto.GammaDto;
 import org.dpsoftware.managers.dto.StateDto;
 import org.dpsoftware.utilities.CommonUtility;
 
 import java.io.IOException;
-import java.text.DecimalFormat;
-import java.time.LocalTime;
-import java.time.temporal.ChronoUnit;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Optional;
@@ -72,20 +62,17 @@ import java.util.Optional;
 @Slf4j
 public class SettingsController {
 
+    // Inject Tab controllers
+    @FXML private MqttTabController mqttTabController;
+    @FXML private DevicesTabController devicesTabController;
+    @FXML private ModeTabController modeTabController;
+    @FXML private MiscTabController miscTabController;
+    // FXML binding
     @FXML private TabPane mainTabPane;
-
     @FXML private TextField ledStartOffset;
-    @FXML private ComboBox<String> gamma;
-    @FXML private ComboBox<String> audioDevice;
-    @FXML private ComboBox<String> audioChannels;
-    @FXML private Slider audioGain;
-    @FXML private ComboBox<String> whiteTemperature;
-    @FXML private ComboBox<String> effect;
     @FXML private Button saveLedButton;
     @FXML private Button playButton;
-    @FXML private Button saveMiscButton;
     @FXML private Button showTestImageButton;
-    @FXML private CheckBox startWithSystem;
     @FXML private TextField topLed;
     @FXML private TextField leftLed;
     @FXML private TextField rightLed;
@@ -98,32 +85,17 @@ public class SettingsController {
     @FXML private Label version;
     @FXML private final StringProperty producerValue = new SimpleStringProperty("");
     @FXML private final StringProperty consumerValue = new SimpleStringProperty("");
-    @FXML private CheckBox eyeCare;
     @FXML private CheckBox splitBottomRow;
-    @FXML private ComboBox<String> framerate;
-    @FXML private ColorPicker colorPicker;
-    @FXML private ToggleButton toggleLed;
-    @FXML private Slider brightness;
     @FXML private Label bottomLeftLedLabel;
     @FXML private Label bottomRightLedLabel;
     @FXML private Label bottomRowLedLabel;
     @FXML private Label displayLabel;
-    @FXML private Label contextGammaGain;
-    @FXML private Label contextChooseColorChooseLoopback;
-    @FXML private Spinner<LocalTime> nightModeFrom;
-    @FXML private Spinner<LocalTime> nightModeTo;
-    @FXML private Spinner<String> nightModeBrightness;
     ImageView imageView;
     Image controlImage;
     AnimationTimer animationTimer;
     Configuration currentConfig;
     StorageManager sm;
     DisplayManager displayManager;
-    // Inject Tab controllers
-    @FXML private MqttTabController mqttTabController;
-    @FXML private DevicesTabController devicesTabController;
-    @FXML private ModeTabController modeTabController;
-
 
 
     /**
@@ -132,9 +104,11 @@ public class SettingsController {
     @FXML
     protected void initialize() {
 
+        // Inject main controller into children
         mqttTabController.injectSettingsController(this);
         devicesTabController.injectSettingsController(this);
         modeTabController.injectSettingsController(this);
+        miscTabController.injectSettingsController(this);
 
         Platform.setImplicitExit(false);
         sm = new StorageManager();
@@ -149,17 +123,6 @@ public class SettingsController {
         }
         currentConfig = sm.readConfig(false);
         initComboBox();
-        audioDevice.getItems().add(Constants.DEFAULT_AUDIO_OUTPUT);
-        if (FireflyLuciferin.config != null && AudioLoopback.audioDevices.isEmpty()) {
-            AudioUtility audioLoopback = new AudioLoopbackSoftware();
-            for (String device : audioLoopback.getLoopbackDevices().values()) {
-                if (device.contains(Constants.LOOPBACK)) audioDevice.getItems().add(device);
-            }
-        } else {
-            for (String device : AudioLoopback.audioDevices.values()) {
-                if (device.contains(Constants.LOOPBACK)) audioDevice.getItems().add(device);
-            }
-        }
         if (NativeExecutor.isLinux()) {
             producerLabel.textProperty().bind(producerValueProperty());
             consumerLabel.textProperty().bind(consumerValueProperty());
@@ -174,10 +137,6 @@ public class SettingsController {
             version.setText("by Davide Perini (VERSION)".replaceAll("VERSION", FireflyLuciferin.version));
         }
         orientation.getItems().addAll(Constants.CLOCKWISE, Constants.ANTICLOCKWISE);
-
-        for (Constants.Framerate fps : Constants.Framerate.values()) {
-            framerate.getItems().add(fps.getFramerate());
-        }
         showTestImageButton.setVisible(currentConfig != null);
         setSaveButtonText();
         // Init default values
@@ -187,7 +146,6 @@ public class SettingsController {
         // Force numeric fields
         setNumericTextField();
         runLater();
-
         initListeners();
         startAnimationTimer();
 
@@ -199,18 +157,7 @@ public class SettingsController {
     void initComboBox() {
 
         modeTabController.initComboBox();
-        for (Constants.Gamma gma : Constants.Gamma.values()) {
-            gamma.getItems().add(gma.getGamma());
-        }
-        for (Constants.Effect ef : Constants.Effect.values()) {
-            effect.getItems().add(ef.getEffect());
-        }
-        for (Constants.WhiteTemperature kelvin : Constants.WhiteTemperature.values()) {
-            whiteTemperature.getItems().add(kelvin.getWhiteTemperature());
-        }
-        for (Constants.AudioChannels audioChan : Constants.AudioChannels.values()) {
-            audioChannels.getItems().add(audioChan.getAudioChannels());
-        }
+        miscTabController.initComboBox();
 
     }
 
@@ -225,10 +172,7 @@ public class SettingsController {
             mqttTabController.initDefaultValues();
             devicesTabController.initDefaultValues();
             modeTabController.initDefaultValues();
-            gamma.setValue(Constants.GAMMA_DEFAULT);
-            whiteTemperature.setValue(Constants.WhiteTemperature.UNCORRECTEDTEMPERATURE.getWhiteTemperature());
-            effect.setValue(Constants.Effect.BIAS_LIGHT.getEffect());
-            framerate.setValue("30 FPS");
+            miscTabController.initDefaultValues();
             orientation.setValue(Constants.CLOCKWISE);
             topLed.setText("33");
             leftLed.setText("18");
@@ -236,8 +180,6 @@ public class SettingsController {
             bottomLeftLed.setText("13");
             bottomRightLed.setText("13");
             bottomRowLed.setText("26");
-            toggleLed.setSelected(true);
-            brightness.setValue(255);
             bottomLeftLed.setVisible(true);
             bottomRightLed.setVisible(true);
             bottomRowLed.setVisible(false);
@@ -245,23 +187,11 @@ public class SettingsController {
             bottomRightLedLabel.setVisible(true);
             bottomRowLedLabel.setVisible(false);
             splitBottomRow.setSelected(true);
-            audioGain.setVisible(false);
-            audioDevice.setVisible(false);
-            audioChannels.setVisible(false);
-            audioChannels.setValue(Constants.AudioChannels.AUDIO_CHANNEL_2.getAudioChannels());
-            audioDevice.setValue(Constants.DEFAULT_AUDIO_OUTPUT);
-            WidgetFactory widgetFactory = new WidgetFactory();
-            nightModeFrom.setValueFactory(widgetFactory.timeSpinnerValueFactory(LocalTime.now().withHour(22).withMinute(0).truncatedTo(ChronoUnit.MINUTES)));
-            nightModeTo.setValueFactory(widgetFactory.timeSpinnerValueFactory(LocalTime.now().withHour(8).withMinute(0).truncatedTo(ChronoUnit.MINUTES)));
-            nightModeBrightness.setValueFactory(widgetFactory.spinnerNightModeValueFactory());
-            enableDisableNightMode(Constants.NIGHT_MODE_OFF);
         } else {
             initValuesFromSettingsFile();
         }
 
     }
-
-
 
     /**
      * Init form values by reading existing config file
@@ -285,9 +215,7 @@ public class SettingsController {
                 break;
             case 3: displayLabel.setText(Constants.LEFT_DISPLAY); break;
         }
-        if (NativeExecutor.isWindows()) {
-            startWithSystem.setSelected(currentConfig.isStartWithSystem());
-        } else if (FireflyLuciferin.config.isToggleLed() && (Constants.Effect.BIAS_LIGHT.getEffect().equals(FireflyLuciferin.config.getEffect())
+        if (!NativeExecutor.isWindows() && FireflyLuciferin.config.isToggleLed() && (Constants.Effect.BIAS_LIGHT.getEffect().equals(FireflyLuciferin.config.getEffect())
                 || Constants.Effect.MUSIC_MODE_VU_METER.getEffect().equals(FireflyLuciferin.config.getEffect())
                 || Constants.Effect.MUSIC_MODE_BRIGHT.getEffect().equals(FireflyLuciferin.config.getEffect())
                 || Constants.Effect.MUSIC_MODE_RAINBOW.getEffect().equals(FireflyLuciferin.config.getEffect()))) {
@@ -295,10 +223,6 @@ public class SettingsController {
             setButtonImage();
         }
         ledStartOffset.setText(String.valueOf(currentConfig.getLedStartOffset()));
-        gamma.setValue(String.valueOf(currentConfig.getGamma()));
-        whiteTemperature.setValue(Constants.WhiteTemperature.values()[currentConfig.getWhiteTemperature()-1].getWhiteTemperature());
-        framerate.setValue(currentConfig.getDesiredFramerate() + ((currentConfig.getDesiredFramerate().equals(Constants.UNLOCKED)) ? "" : " FPS"));
-        eyeCare.setSelected(currentConfig.isEyeCare());
         orientation.setValue(currentConfig.getOrientation());
         topLed.setText(String.valueOf(currentConfig.getTopLed()));
         leftLed.setText(String.valueOf(currentConfig.getLeftLed()));
@@ -306,58 +230,13 @@ public class SettingsController {
         bottomLeftLed.setText(String.valueOf(currentConfig.getBottomLeftLed()));
         bottomRightLed.setText(String.valueOf(currentConfig.getBottomRightLed()));
         bottomRowLed.setText(String.valueOf(currentConfig.getBottomRowLed()));
-        String[] color = (FireflyLuciferin.config.getColorChooser().equals(Constants.DEFAULT_COLOR_CHOOSER)) ?
-                currentConfig.getColorChooser().split(",") : FireflyLuciferin.config.getColorChooser().split(",");
-        colorPicker.setValue(Color.rgb(Integer.parseInt(color[0]), Integer.parseInt(color[1]), Integer.parseInt(color[2]), Double.parseDouble(color[3])/255));
-        brightness.setValue((Double.parseDouble(color[3])/255)*100);
-        audioGain.setValue(currentConfig.getAudioLoopbackGain());
-        audioChannels.setValue(currentConfig.getAudioChannels());
-        audioDevice.setValue(currentConfig.getAudioDevice());
-        effect.setValue(FireflyLuciferin.config.getEffect());
-        if (FireflyLuciferin.config.isToggleLed()) {
-            toggleLed.setText(Constants.TURN_LED_OFF);
-        } else {
-            toggleLed.setText(Constants.TURN_LED_ON);
-        }
-        toggleLed.setSelected(FireflyLuciferin.config.isToggleLed());
         splitBottomRow.setSelected(currentConfig.isSplitBottomRow());
-        WidgetFactory widgetFactory = new WidgetFactory();
-        nightModeFrom.setValueFactory(widgetFactory.timeSpinnerValueFactory(FireflyLuciferin.config.getNightModeFrom()));
-        nightModeTo.setValueFactory(widgetFactory.timeSpinnerValueFactory(FireflyLuciferin.config.getNightModeTo()));
-        nightModeBrightness.setValueFactory(widgetFactory.spinnerNightModeValueFactory());
-        enableDisableNightMode(nightModeBrightness.getValue());
         mqttTabController.initValuesFromSettingsFile(currentConfig);
         devicesTabController.initValuesFromSettingsFile(currentConfig);
         modeTabController.initValuesFromSettingsFile(currentConfig);
+        miscTabController.initValuesFromSettingsFile(currentConfig);
         splitBottomRow();
-        setContextMenu();
-
-    }
-
-    /**
-     * Setup the context menu based on the selected effect
-     */
-    void setContextMenu() {
-
-        if (Constants.Effect.MUSIC_MODE_VU_METER.getEffect().equals(FireflyLuciferin.config.getEffect())
-                || Constants.Effect.MUSIC_MODE_BRIGHT.getEffect().equals(FireflyLuciferin.config.getEffect())
-                || Constants.Effect.MUSIC_MODE_RAINBOW.getEffect().equals(FireflyLuciferin.config.getEffect()))  {
-            colorPicker.setVisible(false);
-            contextChooseColorChooseLoopback.setText(Constants.CONTEXT_MENU_AUDIO_DEVICE);
-            gamma.setVisible(false);
-            contextGammaGain.setText(Constants.CONTEXT_MENU_AUDIO_GAIN);
-            audioGain.setVisible(true);
-            audioDevice.setVisible(true);
-            audioChannels.setVisible(true);
-        } else {
-            colorPicker.setVisible(true);
-            contextChooseColorChooseLoopback.setText(Constants.CONTEXT_MENU_COLOR);
-            gamma.setVisible(true);
-            contextGammaGain.setText(Constants.CONTEXT_MENU_GAMMA);
-            audioGain.setVisible(false);
-            audioDevice.setVisible(false);
-            audioChannels.setVisible(false);
-        }
+        miscTabController.setContextMenu();
 
     }
 
@@ -380,23 +259,6 @@ public class SettingsController {
             devicesTabController.setTableEdit();
             orientation.requestFocus();
         });
-
-    }
-
-    /**
-     * Send serialParams, this will cause a reboot on the microcontroller
-     */
-    void sendSerialParams() {
-
-        java.awt.Color[] leds = new java.awt.Color[1];
-        try {
-            leds[0] = new java.awt.Color((int)(colorPicker.getValue().getRed() * 255),
-                    (int)(colorPicker.getValue().getGreen() * 255),
-                    (int)(colorPicker.getValue().getBlue() * 255));
-            FireflyLuciferin.sendColorsViaUSB(leds);
-        } catch (IOException e) {
-            log.error(e.getMessage());
-        }
 
     }
 
@@ -435,79 +297,11 @@ public class SettingsController {
      */
     private void initListeners() {
 
-        // Toggle LED button listener
-        toggleLed.setOnAction(e -> {
-            if ((toggleLed.isSelected())) {
-                toggleLed.setText(Constants.TURN_LED_OFF);
-                turnOnLEDs(currentConfig, true);
-                if (FireflyLuciferin.config != null) {
-                    FireflyLuciferin.config.setToggleLed(true);
-                }
-            } else {
-                toggleLed.setText(Constants.TURN_LED_ON);
-                turnOffLEDs(currentConfig);
-                if (FireflyLuciferin.config != null) {
-                    FireflyLuciferin.config.setToggleLed(false);
-                }
-            }
-        });
-        // Color picker listener
-        EventHandler<ActionEvent> colorPickerEvent = e -> turnOnLEDs(currentConfig, true);
-        colorPicker.setOnAction(colorPickerEvent);
-        // Gamma can be changed on the fly
-        gamma.valueProperty().addListener((ov, t, gamma) -> {
-            if (currentConfig != null && currentConfig.isMqttEnable()) {
-                GammaDto gammaDto = new GammaDto();
-                gammaDto.setGamma(Double.parseDouble(gamma));
-                MQTTManager.publishToTopic(MQTTManager.getMqttTopic(Constants.MQTT_GAMMA),
-                        CommonUtility.writeValueAsString(gammaDto));
-            }
-            FireflyLuciferin.config.setGamma(Double.parseDouble(gamma));
-        });
-        // White temperature can be changed on the fly
-        whiteTemperature.valueProperty().addListener((ov, t, kelvin) -> {
-            FireflyLuciferin.whiteTemperature = whiteTemperature.getSelectionModel().getSelectedIndex() + 1;
-            if (currentConfig != null && currentConfig.isMqttEnable()) {
-                StateDto stateDto = new StateDto();
-                stateDto.setState(Constants.ON);
-                if (!(currentConfig.isMqttEnable() && FireflyLuciferin.RUNNING)) {
-                    stateDto.setEffect(Constants.SOLID);
-                }
-                stateDto.setWhitetemp(FireflyLuciferin.whiteTemperature);
-                MQTTManager.publishToTopic(MQTTManager.getMqttTopic(Constants.MQTT_SET), CommonUtility.writeValueAsString(stateDto));
-            }
-        });
         setSerialPortAvailableCombo();
-        brightness.valueProperty().addListener((ov, oldVal, newVal) -> turnOnLEDs(currentConfig, false));
-        audioGain.valueProperty().addListener((ov, oldVal, newVal) -> {
-            DecimalFormat df = new DecimalFormat(Constants.NUMBER_FORMAT);
-            float selectedGain = Float.parseFloat(df.format(newVal).replace(",","."));
-            FireflyLuciferin.config.setAudioLoopbackGain(selectedGain);
-        });
         splitBottomRow.setOnAction(e -> splitBottomRow());
         mqttTabController.initListeners();
         modeTabController.initListeners();
-        effect.valueProperty().addListener((ov, oldVal, newVal) -> {
-            if (FireflyLuciferin.config != null) {
-                FireflyLuciferin.config.setEffect(newVal);
-                setContextMenu();
-                if (!oldVal.equals(newVal)) {
-                    FireflyLuciferin.guiManager.stopCapturingThreads(true);
-                    CommonUtility.sleepMilliseconds(100);
-                    FireflyLuciferin.config.setEffect(newVal);
-                    FireflyLuciferin.config.setToggleLed(true);
-                    turnOnLEDs(currentConfig, true);
-                }
-            }
-        });
-        nightModeFrom.valueProperty().addListener((obs, oldValue, newValue) -> FireflyLuciferin.config.setNightModeFrom(newValue));
-        nightModeTo.valueProperty().addListener((obs, oldValue, newValue) -> FireflyLuciferin.config.setNightModeTo(newValue));
-        nightModeBrightness.valueProperty().addListener((obs, oldValue, newValue) -> {
-            if (FireflyLuciferin.config != null) {
-                FireflyLuciferin.config.setNightModeBrightness(newValue);
-            }
-            enableDisableNightMode(newValue);
-        });
+        miscTabController.initListeners(currentConfig);
         devicesTabController.multiMonitor.valueProperty().addListener((ov, t, value) -> {
             if (!modeTabController.serialPort.isFocused()) {
                 if (!value.equals(Constants.MULTIMONITOR_1)) {
@@ -533,6 +327,7 @@ public class SettingsController {
      * Add bold style to the available serial ports
      */
     void setSerialPortAvailableCombo() {
+
         Map<String, Boolean> availableDevices = FireflyLuciferin.getAvailableDevices();
         modeTabController.serialPort.setCellFactory(new Callback<>() {
             @Override
@@ -559,21 +354,6 @@ public class SettingsController {
                 };
             }
         });
-    }
-
-    /**
-     * Toggle night mode params
-     * @param nightModeBrightness brightness param for night mode
-     */
-    void enableDisableNightMode(String nightModeBrightness) {
-
-        if (nightModeBrightness.equals(Constants.NIGHT_MODE_OFF)) {
-            nightModeFrom.setDisable(true);
-            nightModeTo.setDisable(true);
-        } else {
-            nightModeFrom.setDisable(false);
-            nightModeTo.setDisable(false);
-        }
 
     }
 
@@ -600,26 +380,26 @@ public class SettingsController {
             saveLedButton.setText(Constants.SAVE);
             modeTabController.saveSettingsButton.setText(Constants.SAVE);
             mqttTabController.saveMQTTButton.setText(Constants.SAVE);
-            saveMiscButton.setText(Constants.SAVE);
+            miscTabController.saveMiscButton.setText(Constants.SAVE);
             devicesTabController.saveDeviceButton.setText(Constants.SAVE);
             if (NativeExecutor.isWindows()) {
                 saveLedButton.setPrefWidth(95);
                 modeTabController.saveSettingsButton.setPrefWidth(95);
                 mqttTabController.saveMQTTButton.setPrefWidth(95);
-                saveMiscButton.setPrefWidth(95);
+                miscTabController.saveMiscButton.setPrefWidth(95);
                 devicesTabController.saveDeviceButton.setPrefWidth(95);
             } else {
                 saveLedButton.setPrefWidth(125);
                 modeTabController.saveSettingsButton.setPrefWidth(125);
                 mqttTabController.saveMQTTButton.setPrefWidth(125);
-                saveMiscButton.setPrefWidth(125);
+                miscTabController.saveMiscButton.setPrefWidth(125);
                 devicesTabController.saveDeviceButton.setPrefWidth(125);
             }
         } else {
             saveLedButton.setText(Constants.SAVE_AND_CLOSE);
             modeTabController.saveSettingsButton.setText(Constants.SAVE_AND_CLOSE);
             mqttTabController.saveMQTTButton.setText(Constants.SAVE_AND_CLOSE);
-            saveMiscButton.setText(Constants.SAVE_AND_CLOSE);
+            miscTabController.saveMiscButton.setText(Constants.SAVE_AND_CLOSE);
             devicesTabController.saveDeviceButton.setText(Constants.SAVE_AND_CLOSE);
         }
 
@@ -629,6 +409,7 @@ public class SettingsController {
      * Show hide bottom row options
      */
     private void splitBottomRow() {
+
         if (splitBottomRow.isSelected()) {
             bottomLeftLed.setVisible(true);
             bottomRightLed.setVisible(true);
@@ -644,6 +425,7 @@ public class SettingsController {
             bottomRightLedLabel.setVisible(false);
             bottomRowLedLabel.setVisible(true);
         }
+
     }
 
     /**
@@ -677,8 +459,8 @@ public class SettingsController {
             config.setScreenResY(Integer.parseInt(modeTabController.screenHeight.getText()));
             config.setLedStartOffset(Integer.parseInt(ledStartOffset.getText()));
             config.setOsScaling(Integer.parseInt((modeTabController.scaling.getValue()).replace(Constants.PERCENT,"")));
-            config.setGamma(Double.parseDouble(gamma.getValue()));
-            config.setWhiteTemperature(whiteTemperature.getSelectionModel().getSelectedIndex() + 1);
+            config.setGamma(Double.parseDouble(miscTabController.gamma.getValue()));
+            config.setWhiteTemperature(miscTabController.whiteTemperature.getSelectionModel().getSelectedIndex() + 1);
             config.setSerialPort(modeTabController.serialPort.getValue());
             config.setDefaultLedMatrix(modeTabController.aspectRatio.getValue().equals(Constants.AUTO_DETECT_BLACK_BARS) ?
                     Constants.AspectRatio.FULLSCREEN.getAspectRatio() : modeTabController.aspectRatio.getValue());
@@ -689,28 +471,28 @@ public class SettingsController {
                 default -> config.setMultiMonitor(1);
             }
             config.setMonitorNumber(modeTabController.monitorNumber.getValue());
-            config.setDesiredFramerate(framerate.getValue().equals(Constants.UNLOCKED) ?
-                    framerate.getValue() : framerate.getValue().split(" ")[0]);
+            config.setDesiredFramerate(miscTabController.framerate.getValue().equals(Constants.UNLOCKED) ?
+                    miscTabController.framerate.getValue() : miscTabController.framerate.getValue().split(" ")[0]);
             config.setMqttServer(mqttTabController.mqttHost.getText() + ":" + mqttTabController.mqttPort.getText());
             config.setMqttTopic(mqttTabController.mqttTopic.getText());
             config.setMqttUsername(mqttTabController.mqttUser.getText());
             config.setMqttPwd(mqttTabController.mqttPwd.getText());
             config.setMqttEnable(mqttTabController.mqttEnable.isSelected());
-            config.setEyeCare(eyeCare.isSelected());
+            config.setEyeCare(miscTabController.eyeCare.isSelected());
             config.setMqttStream(mqttTabController.mqttStream.isSelected());
             config.setCheckForUpdates(devicesTabController.checkForUpdates.isSelected());
             config.setSyncCheck(devicesTabController.syncCheck.isSelected());
-            config.setToggleLed(toggleLed.isSelected());
-            config.setNightModeFrom(nightModeFrom.getValue());
-            config.setNightModeTo(nightModeTo.getValue());
-            config.setNightModeBrightness(nightModeBrightness.getValue());
-            config.setBrightness((int) (brightness.getValue()/100 *255));
-            config.setAudioChannels(audioChannels.getValue());
-            config.setAudioLoopbackGain((float) audioGain.getValue());
-            config.setAudioDevice(audioDevice.getValue());
-            config.setEffect(effect.getValue());
-            config.setColorChooser((int)(colorPicker.getValue().getRed()*255) + "," + (int)(colorPicker.getValue().getGreen()*255) + ","
-                    + (int)(colorPicker.getValue().getBlue()*255) + "," + (int)(colorPicker.getValue().getOpacity()*255));
+            config.setToggleLed(miscTabController.toggleLed.isSelected());
+            config.setNightModeFrom(miscTabController.nightModeFrom.getValue());
+            config.setNightModeTo(miscTabController.nightModeTo.getValue());
+            config.setNightModeBrightness(miscTabController.nightModeBrightness.getValue());
+            config.setBrightness((int) (miscTabController.brightness.getValue()/100 *255));
+            config.setAudioChannels(miscTabController.audioChannels.getValue());
+            config.setAudioLoopbackGain((float) miscTabController.audioGain.getValue());
+            config.setAudioDevice(miscTabController.audioDevice.getValue());
+            config.setEffect(miscTabController.effect.getValue());
+            config.setColorChooser((int)(miscTabController.colorPicker.getValue().getRed()*255) + "," + (int)(miscTabController.colorPicker.getValue().getGreen()*255) + ","
+                    + (int)(miscTabController.colorPicker.getValue().getBlue()*255) + "," + (int)(miscTabController.colorPicker.getValue().getOpacity()*255));
             if (JavaFXStarter.whoAmI != 1) {
                 Configuration mainConfig = sm.readConfig(true);
                 mainConfig.setGamma(config.getGamma());
@@ -780,12 +562,12 @@ public class SettingsController {
                 case WinAPI -> config.setCaptureMethod(Configuration.CaptureMethod.WinAPI.name());
                 case CPU -> config.setCaptureMethod(Configuration.CaptureMethod.CPU.name());
             }
-            if (startWithSystem.isSelected()) {
+            if (miscTabController.startWithSystem.isSelected()) {
                 nativeExecutor.writeRegistryKey();
             } else {
                 nativeExecutor.deleteRegistryKey();
             }
-            config.setStartWithSystem(startWithSystem.isSelected());
+            config.setStartWithSystem(miscTabController.startWithSystem.isSelected());
         } else if (NativeExecutor.isMac()) {
             if (modeTabController.captureMethod.getValue() == Configuration.CaptureMethod.AVFVIDEOSRC) {
                 config.setCaptureMethod(Configuration.CaptureMethod.AVFVIDEOSRC.name());
@@ -811,11 +593,11 @@ public class SettingsController {
 
         FirmwareConfigDto firmwareConfigDto = new FirmwareConfigDto();
         if (currentConfig.isMqttEnable()) {
-            if (devicesTabController.deviceTableData != null && devicesTabController.deviceTableData.size() > 0) {
+            if (DevicesTabController.deviceTableData != null && DevicesTabController.deviceTableData.size() > 0) {
                 if (Constants.SERIAL_PORT_AUTO.equals(modeTabController.serialPort.getValue())) {
-                    firmwareConfigDto.setMAC(devicesTabController.deviceTableData.get(0).getMac());
+                    firmwareConfigDto.setMAC(DevicesTabController.deviceTableData.get(0).getMac());
                 }
-                devicesTabController.deviceTableData.forEach(glowWormDevice -> {
+                DevicesTabController.deviceTableData.forEach(glowWormDevice -> {
                     if (glowWormDevice.getDeviceName().equals(modeTabController.serialPort.getValue()) || glowWormDevice.getDeviceIP().equals(modeTabController.serialPort.getValue())) {
                         firmwareConfigDto.setMAC(glowWormDevice.getMac());
                     }
@@ -838,7 +620,7 @@ public class SettingsController {
                     MQTTManager.publishToTopic(MQTTManager.getMqttTopic(Constants.MQTT_FIRMWARE_CONFIG), CommonUtility.writeValueAsString(firmwareConfigDto));
                 } else {
                     FireflyLuciferin.baudRate = Constants.BaudRate.valueOf(Constants.BAUD_RATE_PLACEHOLDER + modeTabController.baudRate.getValue()).ordinal() + 1;
-                    sendSerialParams();
+                    miscTabController.sendSerialParams();
                 }
                 exit(e);
             } else if (button == ButtonType.CANCEL) {
@@ -908,7 +690,7 @@ public class SettingsController {
         otherConfig.setBrightness(config.getBrightness());
         otherConfig.setEffect(config.getEffect());
         if (NativeExecutor.isWindows()) {
-            otherConfig.setStartWithSystem(startWithSystem.isSelected());
+            otherConfig.setStartWithSystem(miscTabController.startWithSystem.isSelected());
         }
 
     }
@@ -1057,52 +839,6 @@ public class SettingsController {
     /**
      * Turn ON LEDs
      * @param currentConfig stored config
-     * @param setBrightness brightness level
-     */
-    void turnOnLEDs(Configuration currentConfig, boolean setBrightness) {
-
-        if (setBrightness) {
-            brightness.setValue((int)(colorPicker.getValue().getOpacity()*100));
-        } else {
-            colorPicker.setValue(Color.rgb((int)(colorPicker.getValue().getRed() * 255), (int)(colorPicker.getValue().getGreen() * 255),
-                    (int)(colorPicker.getValue().getBlue() * 255), (brightness.getValue()/100)));
-        }
-        if (currentConfig != null) {
-            if (toggleLed.isSelected() || !setBrightness) {
-                CommonUtility.sleepMilliseconds(100);
-                if (!FireflyLuciferin.RUNNING && (effect.getValue().equals(Constants.Effect.BIAS_LIGHT.getEffect())
-                        || effect.getValue().equals(Constants.Effect.MUSIC_MODE_VU_METER.getEffect())
-                        || effect.getValue().equals(Constants.Effect.MUSIC_MODE_BRIGHT.getEffect())
-                        || effect.getValue().equals(Constants.Effect.MUSIC_MODE_RAINBOW.getEffect()))) {
-                    FireflyLuciferin.guiManager.startCapturingThreads();
-                } else {
-                    if (currentConfig.isMqttEnable()) {
-                        StateDto stateDto = new StateDto();
-                        stateDto.setState(Constants.ON);
-                        if (!FireflyLuciferin.RUNNING) {
-                            stateDto.setEffect(effect.getValue().toLowerCase());
-                        }
-                        ColorDto colorDto = new ColorDto();
-                        colorDto.setR((int)(colorPicker.getValue().getRed() * 255));
-                        colorDto.setG((int)(colorPicker.getValue().getGreen() * 255));
-                        colorDto.setB((int)(colorPicker.getValue().getBlue() * 255));
-                        stateDto.setColor(colorDto);
-                        stateDto.setBrightness(CommonUtility.getNightBrightness());
-                        stateDto.setWhitetemp(FireflyLuciferin.config.getWhiteTemperature());
-                        MQTTManager.publishToTopic(MQTTManager.getMqttTopic(Constants.MQTT_SET), CommonUtility.writeValueAsString(stateDto));
-                    } else {
-                        sendSerialParams();
-                    }
-                    FireflyLuciferin.config.setBrightness((int)((brightness.getValue() / 100) * 255));
-                }
-            }
-        }
-
-    }
-
-    /**
-     * Turn ON LEDs
-     * @param currentConfig stored config
      */
     void turnOffLEDs(Configuration currentConfig) {
 
@@ -1157,40 +893,22 @@ public class SettingsController {
         bottomRightLed.setTooltip(createTooltip(Constants.TOOLTIP_BOTTOMRIGHTLED));
         bottomRowLed.setTooltip(createTooltip(Constants.TOOLTIP_BOTTOMROWLED));
         orientation.setTooltip(createTooltip(Constants.TOOLTIP_ORIENTATION));
-
         ledStartOffset.setTooltip(createTooltip(Constants.TOOLTIP_LEDSTARTOFFSET));
-        gamma.setTooltip(createTooltip(Constants.TOOLTIP_GAMMA));
-        whiteTemperature.setTooltip(createTooltip(Constants.TOOLTIP_WHITE_TEMP));
-        if (NativeExecutor.isWindows()) {
-            startWithSystem.setTooltip(createTooltip(Constants.TOOLTIP_START_WITH_SYSTEM));
-        }
-        framerate.setTooltip(createTooltip(Constants.TOOLTIP_FRAMERATE));
-        eyeCare.setTooltip(createTooltip(Constants.TOOLTIP_EYE_CARE));
-        brightness.setTooltip(createTooltip(Constants.TOOLTIP_BRIGHTNESS));
         splitBottomRow.setTooltip(createTooltip(Constants.TOOLTIP_SPLIT_BOTTOM_ROW));
-        audioDevice.setTooltip(createTooltip(Constants.TOOLTIP_AUDIO_DEVICE));
-        audioChannels.setTooltip(createTooltip(Constants.TOOLTIP_AUDIO_CHANNELS));
-        audioGain.setTooltip(createTooltip(Constants.TOOLTIP_AUDIO_GAIN));
-        effect.setTooltip(createTooltip(Constants.TOOLTIP_EFFECT));
-        colorPicker.setTooltip(createTooltip(Constants.TOOLTIP_COLORS));
-        nightModeFrom.setTooltip(createTooltip(Constants.TOOLTIP_NIGHT_MODE_FROM));
-        nightModeTo.setTooltip(createTooltip(Constants.TOOLTIP_NIGHT_MODE_TO));
-        nightModeBrightness.setTooltip(createTooltip(Constants.TOOLTIP_NIGHT_MODE_BRIGHT));
         mqttTabController.setTooltips(currentConfig);
         devicesTabController.setTooltips(currentConfig);
         modeTabController.setTooltips(currentConfig);
+        miscTabController.setTooltips(currentConfig);
         if (currentConfig == null) {
             if (!NativeExecutor.isWindows()) {
                 playButton.setTooltip(createTooltip(Constants.TOOLTIP_PLAYBUTTON_NULL, 50, 6000));
             }
             saveLedButton.setTooltip(createTooltip(Constants.TOOLTIP_SAVELEDBUTTON_NULL));
-            saveMiscButton.setTooltip(createTooltip(Constants.TOOLTIP_SAVEMQTTBUTTON_NULL));
         } else {
             if (!NativeExecutor.isWindows()) {
                 playButton.setTooltip(createTooltip(Constants.TOOLTIP_PLAYBUTTON, 200, 6000));
             }
             saveLedButton.setTooltip(createTooltip(Constants.TOOLTIP_SAVELEDBUTTON,200, 6000));
-            saveMiscButton.setTooltip(createTooltip(Constants.TOOLTIP_SAVEMQTTBUTTON,200, 6000));
             showTestImageButton.setTooltip(createTooltip(Constants.TOOLTIP_SHOWTESTIMAGEBUTTON,200, 6000));
         }
 
@@ -1343,6 +1061,15 @@ public class SettingsController {
             }
         }
         return new Image(this.getClass().getResource(imgPath).toString(), true);
+
+    }
+
+    /**
+     * Send serial params
+     */
+    public void sendSerialParams() {
+
+        miscTabController.sendSerialParams();
 
     }
