+++ src/main/resources/org/dpsoftware/gui/linuxSettings.fxml
@@ -281,84 +281,7 @@
       <fx:include fx:id="mqttTab" source="mqttTab.fxml"/>
    </Tab>
    <Tab text="Misc">
-      <content>
-         <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0" />
-
-         <GridPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="400.0">
-            <columnConstraints>
-               <ColumnConstraints hgrow="SOMETIMES" maxWidth="295.0" minWidth="10.0" prefWidth="98.0" />
-               <ColumnConstraints hgrow="SOMETIMES" maxWidth="458.00001525878906" minWidth="10.0" prefWidth="202.66666666666669" />
-            </columnConstraints>
-            <rowConstraints>
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="40.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="40.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="40.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="40.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="40.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="40.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="40.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="40.0" vgrow="SOMETIMES" />
-            </rowConstraints>
-            <padding>
-               <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
-            </padding>
-            <children>
-               <Label text="LEDs control" />
-               <Label fx:id="contextChooseColorChooseLoopback" text="Choose color" GridPane.rowIndex="1" />
-               <Label text="Eye care" GridPane.rowIndex="4" />
-               <Label text="Capture framerate" GridPane.rowIndex="6" />
-               <Label wrapText="true" GridPane.rowIndex="6" />
-               <Button fx:id="saveMiscButton" mnemonicParsing="false" onMouseClicked="#save" prefHeight="25.0" prefWidth="130.0" text="Save and close" GridPane.columnIndex="1" GridPane.rowIndex="7">
-                  <GridPane.margin>
-                     <Insets left="103.0" top="-5.0" />
-                  </GridPane.margin>
-               </Button>
-               <ColorPicker fx:id="colorPicker" prefHeight="25.0" prefWidth="289.0" GridPane.columnIndex="1" GridPane.rowIndex="1">
-                  <value>
-                     <Color blue="1.0" />
-                  </value>
-               </ColorPicker>
-               <ToggleButton fx:id="toggleLed" mnemonicParsing="false" prefHeight="26.0" prefWidth="110.0" text="Turn LEDs ON" GridPane.columnIndex="1" GridPane.rowIndex="0" />
-               <ComboBox fx:id="framerate" prefHeight="25.0" prefWidth="258.0" GridPane.columnIndex="1" GridPane.rowIndex="6" />
-               <Label text="Brightness" GridPane.rowIndex="5" />
-               <Slider fx:id="brightness" blockIncrement="1.0" majorTickUnit="10.0" minorTickCount="5" prefHeight="14.0" prefWidth="239.0" showTickLabels="true" showTickMarks="true" value="100.0" GridPane.columnIndex="1" GridPane.rowIndex="5" />
-               <Label fx:id="contextGammaGain" text="Gamma" GridPane.rowIndex="2" />
-               <ComboBox fx:id="gamma" prefHeight="25.0" prefWidth="465.0" GridPane.columnIndex="1" GridPane.rowIndex="2" />
-               <Label text="White temperature" GridPane.rowIndex="3" />
-               <ComboBox fx:id="whiteTemperature" prefHeight="25.0" prefWidth="459.0" GridPane.columnIndex="1" GridPane.rowIndex="3" />
-               <ComboBox fx:id="effect" prefHeight="25.0" prefWidth="110.0" GridPane.columnIndex="1" GridPane.halignment="RIGHT" />
-               <Slider fx:id="audioGain" blockIncrement="0.5" majorTickUnit="1.0" max="5.0" min="-5.0" minorTickCount="1" showTickLabels="true" showTickMarks="true" snapToTicks="true" GridPane.columnIndex="1" GridPane.rowIndex="2" />
-               <ComboBox fx:id="audioDevice" prefHeight="25.0" prefWidth="110.0" GridPane.columnIndex="1" GridPane.rowIndex="1" />
-               <ComboBox fx:id="audioChannels" prefHeight="25.0" prefWidth="110.0" GridPane.columnIndex="1" GridPane.halignment="RIGHT" GridPane.rowIndex="1" />
-               <HBox prefHeight="100.0" prefWidth="200.0" GridPane.columnIndex="1" GridPane.rowIndex="4">
-                  <children>
-                     <CheckBox fx:id="eyeCare" mnemonicParsing="false">
-                        <HBox.margin>
-                           <Insets right="15.0" top="11.0" />
-                        </HBox.margin>
-                     </CheckBox>
-                     <Spinner fx:id="nightModeFrom" prefHeight="25.0" prefWidth="55.0">
-                        <HBox.margin>
-                           <Insets right="15.0" top="7.0" />
-                        </HBox.margin>
-                     </Spinner>
-                     <Spinner fx:id="nightModeTo" prefHeight="25.0" prefWidth="55.0">
-                        <HBox.margin>
-                           <Insets right="15.0" top="7.0" />
-                        </HBox.margin>
-                     </Spinner>
-                     <Spinner fx:id="nightModeBrightness" prefHeight="25.0" prefWidth="55.0">
-                        <HBox.margin>
-                           <Insets top="7.0" />
-                        </HBox.margin>
-                     </Spinner>
-                  </children>
-               </HBox>
-            </children>
-
-         </GridPane>
-
-      </content>
+      <fx:include fx:id="miscTab" source="miscTab.fxml"/>
    </Tab>
    <Tab text="Devices">
       <fx:include fx:id="devicesTab" source="devicesTab.fxml"/>
