+++ src/main/resources/org/dpsoftware/gui/settings.fxml
@@ -2,11 +2,10 @@
 
 <?import javafx.geometry.*?>
 <?import javafx.scene.control.*?>
-<?import javafx.scene.image.*?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.*?>
-<?import javafx.scene.paint.*?>
-<?import javafx.scene.text.*?>
-
+<?import javafx.scene.text.Font?>
 <TabPane fx:id="mainTabPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="415.0" prefWidth="400.0" stylesheets="@css/main.css" tabClosingPolicy="UNAVAILABLE" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.dpsoftware.gui.SettingsController">
    <Tab text="LEDs config">
       <content>
@@ -207,88 +206,7 @@
       <fx:include fx:id="mqttTab" source="mqttTab.fxml"/>
    </Tab>
    <Tab text="Misc">
-      <content>
-         <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0" />
-
-         <GridPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="400.0">
-            <columnConstraints>
-               <ColumnConstraints hgrow="SOMETIMES" maxWidth="295.0" minWidth="10.0" prefWidth="98.0" />
-               <ColumnConstraints hgrow="SOMETIMES" maxWidth="458.00001525878906" minWidth="10.0" prefWidth="202.66666666666669" />
-            </columnConstraints>
-            <rowConstraints>
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="39.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="39.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="39.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="39.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="39.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="39.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="39.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="39.0" vgrow="SOMETIMES" />
-               <RowConstraints maxHeight="242.0" minHeight="10.0" prefHeight="49.0" vgrow="SOMETIMES" />
-            </rowConstraints>
-            <padding>
-               <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
-            </padding>
-            <children>
-               <Label text="LEDs control" />
-               <Label fx:id="contextChooseColorChooseLoopback" text="Choose color" GridPane.rowIndex="1" />
-               <Label text="Run at login" GridPane.rowIndex="4" />
-               <Label text="Eye care" GridPane.rowIndex="5" />
-               <Label text="Capture framerate" GridPane.rowIndex="7" />
-               <Label wrapText="true" GridPane.rowIndex="6" />
-               <Label wrapText="true" GridPane.rowIndex="7" />
-               <Button fx:id="saveMiscButton" mnemonicParsing="false" onMouseClicked="#save" text="Save and close" GridPane.columnIndex="1" GridPane.rowIndex="8">
-                  <GridPane.margin>
-                     <Insets left="138.0" top="-5.0" />
-                  </GridPane.margin>
-               </Button>
-               <ColorPicker fx:id="colorPicker" prefHeight="25.0" prefWidth="289.0" GridPane.columnIndex="1" GridPane.rowIndex="1">
-                  <value>
-                     <Color blue="1.0" />
-                  </value>
-               </ColorPicker>
-               <ToggleButton fx:id="toggleLed" mnemonicParsing="false" prefHeight="25.0" prefWidth="110.0" text="Turn LEDs ON" GridPane.columnIndex="1" GridPane.rowIndex="0" />
-               <CheckBox fx:id="startWithSystem" mnemonicParsing="false" GridPane.columnIndex="1" GridPane.rowIndex="4" />
-               <ComboBox fx:id="framerate" prefHeight="25.0" prefWidth="258.0" GridPane.columnIndex="1" GridPane.rowIndex="7" />
-               <Label text="Brightness" GridPane.rowIndex="6" />
-               <Slider fx:id="brightness" blockIncrement="1.0" majorTickUnit="10.0" minorTickCount="5" prefHeight="37.0" prefWidth="93.0" showTickLabels="true" showTickMarks="true" value="100.0" GridPane.columnIndex="1" GridPane.rowIndex="6" />
-               <Label fx:id="contextGammaGain" text="Gamma" GridPane.rowIndex="2" />
-               <ComboBox fx:id="gamma" prefHeight="25.0" prefWidth="257.0" GridPane.columnIndex="1" GridPane.rowIndex="2" />
-               <Label text="White temperature" GridPane.rowIndex="3" />
-               <ComboBox fx:id="whiteTemperature" prefHeight="25.0" prefWidth="400.0" GridPane.columnIndex="1" GridPane.rowIndex="3" />
-               <ComboBox fx:id="effect" prefHeight="25.0" prefWidth="110.0" GridPane.columnIndex="1" GridPane.halignment="RIGHT" />
-               <Slider fx:id="audioGain" blockIncrement="0.5" majorTickUnit="1.0" max="5.0" min="-5.0" minorTickCount="1" prefHeight="14.0" prefWidth="110.0" showTickLabels="true" showTickMarks="true" snapToTicks="true" GridPane.columnIndex="1" GridPane.rowIndex="2" />
-               <ComboBox fx:id="audioChannels" prefHeight="25.0" prefWidth="110.0" GridPane.columnIndex="1" GridPane.halignment="RIGHT" GridPane.rowIndex="1" />
-               <ComboBox fx:id="audioDevice" prefHeight="25.0" prefWidth="110.0" GridPane.columnIndex="1" GridPane.rowIndex="1" />
-               <HBox prefHeight="100.0" prefWidth="200.0" GridPane.columnIndex="1" GridPane.rowIndex="5">
-                  <children>
-                     <CheckBox fx:id="eyeCare" mnemonicParsing="false">
-                        <HBox.margin>
-                           <Insets right="15.0" top="11.0" />
-                        </HBox.margin>
-                     </CheckBox>
-                     <Spinner fx:id="nightModeFrom" prefHeight="25.0" prefWidth="55.0">
-                        <HBox.margin>
-                           <Insets right="15.0" top="7.0" />
-                        </HBox.margin>
-                     </Spinner>
-                     <Spinner fx:id="nightModeTo" prefHeight="25.0" prefWidth="55.0">
-                        <HBox.margin>
-                           <Insets right="15.0" top="7.0" />
-                        </HBox.margin>
-                     </Spinner>
-                     <Spinner fx:id="nightModeBrightness" prefWidth="55.0">
-                        <HBox.margin>
-                           <Insets top="7.0" />
-                        </HBox.margin>
-                     </Spinner>
-                  </children>
-               </HBox>
-            </children>
-
-         </GridPane>
-
-      </content>
+      <fx:include fx:id="miscTab" source="miscTab.fxml"/>
    </Tab>
    <Tab text="Devices">
       <fx:include fx:id="devicesTab" source="devicesTab.fxml"/>
